move embedlyInfo class to common
